Move views to blueprints, ignore trailing slashes
