+++ templates/catalog-view.html
@@ -55,8 +55,8 @@
         {% endif %}
         <div class="thread border-top rounded">
             {% if thread["media"] %}
-            <a href="{{ url_for("view_thread", thread_id=thread["id"]) }}">
-            <img class="border-left border-right w-100 rounded-top" src="{{ url_for("uploaded_thumb",
+            <a href="{{ url_for("threads.view", thread_id=thread["id"]) }}">
+            <img class="border-left border-right w-100 rounded-top" src="{{ url_for("upload.thumb",
             media_id=thread["media"]) }}">
             </a>
             {% endif %}
@@ -75,7 +75,7 @@
         </div>
         <div class="{{ "pt-1 border border-top-0 rounded-bottom
         " + ns.bg_style }}">
-        <a href="{{ url_for("view_thread", thread_id=thread["id"]) }}" class="btn btn-dark m-2">View thread</a>
+        <a href="{{ url_for("threads.view", thread_id=thread["id"]) }}" class="btn btn-dark m-2">View thread</a>
         <span class="{{ "mt-3 mr-2 float-right " + ns.text_style }}">
         <i class="fas fa-comment"></i> {{ thread["num_replies"] }}
         <i class="fas fa-image"></i> {{ thread["num_media"] }}
